Alx Low Level Programming

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -8,9 +8,9 @@
  */
 void puts_half(char *str)
 {
-	int n;
-	int j = (strlen(str)) / 2;
-	int k = strlen(str);
+	long n;
+	long j = (strlen(str)) / 2;
+	long k = strlen(str);
 
 	for (n = j; n < k; n++)
 		_putchar(str[n]);
